# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Multi-Tenant Offline-First Desk Booking & Pod Allocation API",
    "version": "1.0.0",
    "description": "Enterprise-grade REST API powering multi-tenant workspace management, offline-first SQLite sync,\ninteractive 2D SVG workstation booking, agile team pod allocations, role-scoped bulk cancellations,\nand Centralized vs. Delegated governance architectures.\n",
    "contact": {
      "name": "Platform Engineering Team",
      "email": "support@deskbooking.enterprise.internal"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.deskbooking.enterprise.internal/api",
      "description": "Production Environment"
    }
  ],
...
  "paths": {
...
    "/organizations": {
      "get": {
        "tags": [
          "Organizations & Tenancy"
        ],
        "summary": "List organizations accessible by caller",
        "operationId": "listOrganizations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of tenant organizations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organizations & Tenancy"
        ],
        "summary": "Create new tenant organization",
        "operationId": "createOrganization",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Cyberdyne Systems"
                  },
                  "operatingMode": {
                    "type": "string",
                    "enum": [
                      "DELEGATED",
                      "CENTRALIZED"
                    ],
                    "default": "DELEGATED"
                  },
                  "primaryColor": {
                    "type": "string",
                    "example": "#059669"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{id}": {
      "get": {
        "tags": [
          "Organizations & Tenancy"
        ],
        "summary": "Get organization details and brand configuration",
        "operationId": "getOrganization",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Organizations & Tenancy"
        ],
        "summary": "Update organization settings, operating mode, or branding",
        "operationId": "updateOrganization",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "operatingMode": {
                    "type": "string",
                    "enum": [
                      "DELEGATED",
                      "CENTRALIZED"
                    ]
                  },
                  "primaryColor": {
                    "type": "string",
                    "example": "#7c3aed"
                  },
                  "accentColor": {
                    "type": "string",
                    "example": "#a855f7"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Standard Bearer token authentication header"
      }
    },
    "schemas": {
...
      "Organization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "org-acme-1"
          },
          "name": {
            "type": "string",
            "example": "Acme Corporation"
          },
          "operatingMode": {
            "type": "string",
            "enum": [
              "DELEGATED",
              "CENTRALIZED"
            ],
            "example": "DELEGATED"
          },
          "primaryColor": {
            "type": "string",
            "example": "#059669"
          },
          "accentColor": {
            "type": "string",
            "example": "#10b981"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  }
}